fix: controls width
